Don't crash the device if the host crashes before reading the response
If the host does an interrupt write and then never reads the reply, the ColorHug
device cannot process other commands.

More importantly, it also can't be used with HID, when using Microsoft Windows.
Re-enumerate the USB connection if the host fails to read the command response
within 400ms of the command write, which allows the device to automatically
"unstick" itself.