Increase the integral time when measuring very dark samples
The TAOS sensor we're using just doesn't have the sensitivity to do short
readings when measuring close to zero black.
To detect this case do a short read, and increase the integral time if there is
an output frequency less than about 100Hz.

This commit needs a new version of all the client tools, as we can take up to
3.1 seconds to measure absolute black (X=0.0,Y=0.0,Z=0.0).