firmware: Makefile: Linux build w/ microchip toolchain
This patch adds a Makefile.microchip to the firmware which enables
native builds with Linux. Microchip's proprietary but free-as-in-beer
toolchain and application library are required. A quick introduction
to the requirements and links to the Microchip packages are
in the comments at the start of Makefile.microchip.

After downloading and installing the toolchain and the library
packages, and adding some directories to your PATH, you will
be able to build the firmware by typing

make -j

The Makefile which uses sdcc compilation has been moved to
Makefile.sdcc. An entry point Makefile lets you choose which
toolchain to use. It currently defaults to the Microchip toolchain.

Signed-off-by: Thilo Fromm <github@thilo-fromm.de>
Signed-off-by: Richard Hughes <richard@hughsie.com>